Remove hfrunner PoC component
hfrunner was a proof-of-concept that ran system/container checks at 1s
intervals and routed them into the observer pipeline. It was not ready for
main and is being removed.

- Delete comp/anomalydetection/hfrunner/ entirely
- Remove HFRunner from observer Requires and NewComponent wiring
- Remove hfFilterSources / hfFilteredHandle / sourceProvider (no longer needed)
- Remove hfrunnerfx.Module() from agent run command
- Clean up Bazel deps and test srcs
- Move sampleNoSource/countingHandle helpers to test_helpers_test.go
