--- comp/anomalydetection/hfrunner/def/BUILD.bazel
@@ -1,12 +1,0 @@
-load("@rules_go//go:def.bzl", "go_library")
-
-go_library(
-    name = "def",
-    srcs = ["component.go"],
-    importpath = "github.com/DataDog/datadog-agent/comp/anomalydetection/hfrunner/def",
-    visibility = ["//visibility:public"],
-    deps = [
-        "//comp/anomalydetection/observer/def",
-        "//pkg/metrics",
-    ],
-)
