--- comp/anomalydetection/hfrunner/fx-noop/BUILD.bazel
@@ -1,14 +1,0 @@
-load("@rules_go//go:def.bzl", "go_library")
-
-go_library(
-    name = "fx-noop",
-    srcs = ["fx_noop.go"],
-    importpath = "github.com/DataDog/datadog-agent/comp/anomalydetection/hfrunner/fx-noop",
-    visibility = ["//visibility:public"],
-    deps = [
-        "//comp/anomalydetection/hfrunner/def",
-        "//comp/anomalydetection/observer/def",
-        "//pkg/metrics",
-        "//pkg/util/fxutil",
-    ],
-)
