--- comp/anomalydetection/hfrunner/fx/BUILD.bazel
@@ -1,17 +1,0 @@
-load("@rules_go//go:def.bzl", "go_library")
-
-go_library(
-    name = "fx",
-    srcs = [
-        "fx.go",
-        "fx_noop.go",
-    ],
-    importpath = "github.com/DataDog/datadog-agent/comp/anomalydetection/hfrunner/fx",
-    visibility = ["//visibility:public"],
-    deps = [
-        "//comp/anomalydetection/hfrunner/def",
-        "//comp/anomalydetection/hfrunner/impl",
-        "//pkg/util/fxutil",
-        "@org_uber_go_fx//:fx",
-    ],
-)
