--- comp/anomalydetection/hfrunner/fx/fx.go
@@ -1,26 +1,0 @@
-// Unless explicitly stated otherwise all files in this repository are licensed
-// under the Apache License Version 2.0.
-// This product includes software developed at Datadog (https://www.datadoghq.com/).
-// Copyright 2016-present Datadog, Inc.
-
-// The `python` build tag is used here as a proxy for "full agent, not IoT agent".
-// See comp/anomalydetection/hfrunner/fx/fx_noop.go for the stub used in IoT agent builds.
-
-//go:build python
-
-// Package fx defines the fx options for the hfrunner component.
-package fx
-
-import (
-	hfrunnerdef "github.com/DataDog/datadog-agent/comp/anomalydetection/hfrunner/def"
-	hfrunnerimpl "github.com/DataDog/datadog-agent/comp/anomalydetection/hfrunner/impl"
-	"github.com/DataDog/datadog-agent/pkg/util/fxutil"
-)
-
-// Module defines the fx options for this component.
-func Module() fxutil.Module {
-	return fxutil.Component(
-		fxutil.ProvideComponentConstructor(hfrunnerimpl.NewComponent),
-		fxutil.ProvideOptional[hfrunnerdef.Component](),
-	)
-}
