--- comp/anomalydetection/hfrunner/fx/fx_noop.go
@@ -1,20 +1,0 @@
-// Unless explicitly stated otherwise all files in this repository are licensed
-// under the Apache License Version 2.0.
-// This product includes software developed at Datadog (https://www.datadoghq.com/).
-// Copyright 2016-present Datadog, Inc.
-
-//go:build !python
-
-// Package fx defines the fx options for the hfrunner component.
-package fx
-
-import (
-	"go.uber.org/fx"
-
-	"github.com/DataDog/datadog-agent/pkg/util/fxutil"
-)
-
-// Module is a no-op for builds without Python support (e.g. IoT agent).
-func Module() fxutil.Module {
-	return fxutil.Module{Option: fx.Options()}
-}
