--- comp/anomalydetection/hfrunner/impl-noop/BUILD.bazel
@@ -1,13 +1,0 @@
-load("@rules_go//go:def.bzl", "go_library")
-
-go_library(
-    name = "impl-noop",
-    srcs = ["hfrunner.go"],
-    importpath = "github.com/DataDog/datadog-agent/comp/anomalydetection/hfrunner/impl-noop",
-    visibility = ["//visibility:public"],
-    deps = [
-        "//comp/anomalydetection/hfrunner/def",
-        "//comp/anomalydetection/observer/def",
-        "//pkg/metrics",
-    ],
-)
