--- comp/anomalydetection/hfrunner/impl/BUILD.bazel
@@ -1,27 +1,0 @@
-load("@rules_go//go:def.bzl", "go_library")
-
-# gazelle:exclude runner.go
-# gazelle:exclude sender.go
-
-go_library(
-    name = "impl",
-    srcs = [
-        "hfrunner.go",
-        "runner_base.go",
-        "runner_noop.go",
-    ],
-    importpath = "github.com/DataDog/datadog-agent/comp/anomalydetection/hfrunner/impl",
-    visibility = ["//visibility:public"],
-    deps = [
-        "//comp/anomalydetection/hfrunner/def",
-        "//comp/anomalydetection/observer/def",
-        "//comp/core/config",
-        "//comp/core/tagger/def",
-        "//comp/core/workloadfilter/def",
-        "//comp/core/workloadmeta/def",
-        "//comp/def",
-        "//pkg/metrics",
-        "//pkg/util/log",
-        "//pkg/util/option",
-    ],
-)
