+++ comp/anomalydetection/observer/impl/BUILD.bazel
@@ -39,7 +39,6 @@ go_library(
     importpath = "github.com/DataDog/datadog-agent/comp/anomalydetection/observer/impl",
     visibility = ["//visibility:public"],
     deps = [
-        "//comp/anomalydetection/hfrunner/def",
         "//comp/anomalydetection/observer/def",
         "//comp/anomalydetection/observer/impl/patterns",
         "//comp/anomalydetection/recorder/def",
@@ -49,7 +48,6 @@ go_library(
         "//comp/core/telemetry/def",
         "//comp/core/telemetry/impl/noops",
         "//comp/def",
-        "//pkg/metrics",
         "//pkg/util/log",
         "//pkg/util/option",
     ],
@@ -86,7 +84,6 @@ go_test(
         "series_pair_key_test.go",
         "stateview_test.go",
         "storage_test.go",
-        "system_filter_test.go",
         "telemetry_test.go",
         "test_helpers_test.go",
     ],
@@ -96,7 +93,6 @@ go_test(
         "//comp/anomalydetection/observer/def",
         "//comp/anomalydetection/reporter/def",
         "//comp/core/telemetry/impl/noops",
-        "//pkg/metrics",
         "//pkg/util/log",
         "@com_github_stretchr_testify//assert",
         "@com_github_stretchr_testify//require",
