+++ comp/anomalydetection/observer/impl/observer.go
@@ -18,7 +18,6 @@ import (
 
 	compdef "github.com/DataDog/datadog-agent/comp/def"
 
-	hfrunnerdef "github.com/DataDog/datadog-agent/comp/anomalydetection/hfrunner/def"
 	observerdef "github.com/DataDog/datadog-agent/comp/anomalydetection/observer/def"
 	recorderdef "github.com/DataDog/datadog-agent/comp/anomalydetection/recorder/def"
 	reporterdef "github.com/DataDog/datadog-agent/comp/anomalydetection/reporter/def"
@@ -26,7 +25,6 @@ import (
 	log "github.com/DataDog/datadog-agent/comp/core/log/def"
 	telemetry "github.com/DataDog/datadog-agent/comp/core/telemetry/def"
 	"github.com/DataDog/datadog-agent/comp/core/telemetry/impl/noops"
-	"github.com/DataDog/datadog-agent/pkg/metrics"
 
 	pkglog "github.com/DataDog/datadog-agent/pkg/util/log"
 	"github.com/DataDog/datadog-agent/pkg/util/option"
@@ -48,11 +46,6 @@ type Requires struct {
 	// so it receives advance events independently. StorageConsumer reporters receive
 	// storage for windowed log-rate annotations.
 	Reporters []reporterdef.Reporter `group:"anomalydetection_reporters"`
-
-	// HFRunner runs system and container checks at 1s and routes them into the
-	// observer pipeline. The noop variant (hfrunner/fx-noop) is wired for the
-	// main agent build; the real implementation lands with the algorithm PRs.
-	HFRunner hfrunnerdef.Component
 }
 
 // Provides defines the output of the observer component.
@@ -235,7 +228,6 @@ func NewComponent(deps Requires) Provides {
 		obsCh:                make(chan observation, 1000),
 		telemetryHandler:     th,
 		dropCounter:          th.telemetryCounters[telemetryObsChannelDropped],
-		hfFilterSources:      make(map[metrics.MetricSource]struct{}),
 		ingestMetricsEnabled: !cfg.IsConfigured("anomaly_detection.metrics.enabled") || cfg.GetBool("anomaly_detection.metrics.enabled"),
 	}
 
@@ -283,16 +275,6 @@ func NewComponent(deps Requires) Provides {
 	}
 
 	go obs.run()
-
-	// Start high-frequency check runners. The hfrunner component handles
-	// config-based toggling and lifecycle internally; we only collect the
-	// MetricSource sets it wants suppressed from the "all-metrics" pipeline.
-	for src := range deps.HFRunner.StartSystem(obs.GetHandle(hfrunnerdef.HFSource)) {
-		obs.hfFilterSources[src] = struct{}{}
-	}
-	for src := range deps.HFRunner.StartContainer(obs.GetHandle(hfrunnerdef.HFContainerSource)) {
-		obs.hfFilterSources[src] = struct{}{}
-	}
 
 	// Wire agent-internal logs into the observer via the pkg/util/log tap.
 	// anomaly_detection.logs.enabled is the parent gate; without it,
@@ -351,12 +333,6 @@ type observerImpl struct {
 	// Tagged by source for Prometheus visibility. Complements engine.droppedObs
 	// which tracks drops for live/replay parity analysis.
 	dropCounter telemetry.Counter
-
-	// hfFilterSources is the combined set of MetricSource values to suppress from
-	// the "all-metrics" pipeline when their HF counterpart is active. Populated at
-	// construction time from the MetricSource sets returned by hfrunner.StartSystem
-	// and hfrunner.StartContainer.
-	hfFilterSources map[metrics.MetricSource]struct{}
 
 	// ingestMetricsEnabled gates externally-ingested metrics at the handle
 	// factory. When false, "all-metrics" and HF handles return a wrapper
@@ -570,72 +546,24 @@ func (o *observerImpl) GetHandle(name string) observerdef.Handle {
 }
 
 // innerHandle creates the base handle without any middleware wrapping.
-// When any HF check collection is enabled, the "all-metrics" handle is wrapped
-// with hfFilteredHandle to suppress 15s samples for checks that have a 1s HF
-// counterpart active — the scorer should only see the higher-resolution stream.
-// When anomaly_detection.metrics.enabled=false, the resulting handle is further
-// wrapped with metricDropHandle so external metrics are dropped at the edge,
-// while ObserveLog/ObserveProfile pass through.
+// When anomaly_detection.metrics.enabled=false, the handle is wrapped with
+// metricDropHandle so external metrics are dropped at the edge, while
+// ObserveLog/ObserveProfile pass through.
 func (o *observerImpl) innerHandle(name string) observerdef.Handle {
 	h := &handle{ch: o.obsCh, source: name, dropCounter: o.dropCounter}
 	o.engine.registerHandle(h)
 	var out observerdef.Handle = h
-	if len(o.hfFilterSources) > 0 && name == "all-metrics" {
-		out = &hfFilteredHandle{inner: h, sources: o.hfFilterSources}
-	}
 	if !o.ingestMetricsEnabled {
 		out = &metricDropHandle{inner: out}
 	}
 	return out
 }
 
-// sourceProvider is a structural interface satisfied by *metrics.MetricSample,
-// which carries a MetricSource enum populated by the standard check sender.
-// Using a type assertion (rather than adding GetSource to MetricView) avoids
-// importing pkg/metrics into comp/anomalydetection/observer/def.
-type sourceProvider interface {
-	GetSource() metrics.MetricSource
-}
-
-// hfFilteredHandle wraps a Handle and drops metrics whose source is in the
-// provided sources set, so that 15s pipeline samples do not compete with their
-// 1s HF counterparts in the scorer.
-//
-// Filtering uses a MetricSource enum map lookup via a type assertion to
-// sourceProvider. Samples that do not implement sourceProvider pass through
-// unchanged — absence of metadata is not sufficient grounds to drop.
-type hfFilteredHandle struct {
-	inner   observerdef.Handle
-	sources map[metrics.MetricSource]struct{}
-}
-
-func (f *hfFilteredHandle) ObserveMetric(sample observerdef.MetricView) {
-	_ = f.ObserveMetricAndReportDrop(sample)
-}
-
-func (f *hfFilteredHandle) ObserveMetricAndReportDrop(sample observerdef.MetricView) bool {
-	if sp, ok := sample.(sourceProvider); ok {
-		if _, suppressed := f.sources[sp.GetSource()]; suppressed {
-			return false
-		}
-	}
-	if dr, ok := f.inner.(interface {
-		ObserveMetricAndReportDrop(observerdef.MetricView) bool
-	}); ok {
-		return dr.ObserveMetricAndReportDrop(sample)
-	}
-	f.inner.ObserveMetric(sample)
-	return false
-}
-
-func (f *hfFilteredHandle) ObserveLog(msg observerdef.LogView) { f.inner.ObserveLog(msg) }
-
 // metricDropHandle drops every ObserveMetric call but lets logs and
 // profiles through. Used when anomaly_detection.metrics.enabled=false so
-// external metric sources (DogStatsD, check samplers, HF runners) do not
-// feed the engine. Virtual metrics produced by LogMetricsExtractors
-// during engine.IngestLog are unaffected because they bypass this handle
-// path entirely (they are written directly to storage from the engine).
+// external metric sources (DogStatsD, check samplers) do not feed the engine.
+// Virtual metrics produced by LogMetricsExtractors during engine.IngestLog are
+// unaffected because they bypass this handle path entirely.
 type metricDropHandle struct{ inner observerdef.Handle }
 
 var _ observerdef.Handle = (*metricDropHandle)(nil)
