+++ comp/anomalydetection/observer/impl/test_helpers_test.go
@@ -11,6 +11,24 @@ import (
 	observerdef "github.com/DataDog/datadog-agent/comp/anomalydetection/observer/def"
 )
 
+// sampleNoSource implements MetricView only — no sourceProvider.
+type sampleNoSource struct{ name string }
+
+func (s *sampleNoSource) GetName() string         { return s.name }
+func (s *sampleNoSource) GetValue() float64       { return 0 }
+func (s *sampleNoSource) GetRawTags() []string    { return nil }
+func (s *sampleNoSource) GetTimestampUnix() int64 { return 0 }
+func (s *sampleNoSource) GetSampleRate() float64  { return 1 }
+
+// countingHandle records how many MetricView and LogView observations it receives.
+type countingHandle struct {
+	received    int
+	logReceived int
+}
+
+func (h *countingHandle) ObserveMetric(_ observerdef.MetricView) { h.received++ }
+func (h *countingHandle) ObserveLog(_ observerdef.LogView)       { h.logReceived++ }
+
 // mockLogView implements observer.LogView for testing.
 type mockLogView struct {
 	content     string
